+++ lib/tbib_timezone_offset.dart
@@ -4,3 +4,4 @@ library tbib_timezone_offset;
 
 export 'src/tbib_gen_json_date_time.dart';
 export 'src/tbib_timezone_offset.dart';
+export 'widgets/tbib_picker_date_and_time.dart';