+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+## 1.0.0
+
+- now format date with locale
+- add form for date picker or time picker or both from TBIBDatePickerFormField widget see example
+
 ## 0.0.2
 
-* Change example and readme.
+- Change example and readme.
   
 ## 0.0.1
 
-* init.
+- init.